wined3d: Remove some code that doesn't make sense.
This was introduced by abb11f30cd418c6337a93a26981763b20c059172, but I don't
see how it could ever do something useful. Perhaps that's a lack of
imagination on my part, but the only situation in which device->stateBlock
should be NULL is while creating the initial stateblock, and in that case we
just returned in the line above. Setting the stream frequency of all streams
to 0x01010101 doesn't make sense either.