winex11.drv: Allow explorer to focus other process windows.
When explorer gets a take focus message, it tries to focus the foreground
window, but this doesn't work because set_focus can only focus windows in
the current process. We have to look for the focus window in the foreground
thread, not the current one, or we won't find the other process's windows.

Since the other process may crash at any time, causing its windows to be
destroyed, we also have to ignore the BadWindow error that will occur if that
happens at a critical time.