shell32: Fix an uninitialized memory access in SHGetFileInfoW.
If we are passed SHGFI_PIDL, szFullPath is not initialized. If we're passed
the flags PIDL, USEFILEATTRIBUTES, and an icon flag, we will try to use the
uninitialized filename to get the icon. Since the test claims that the
function should return an icon in this case, and we cannot always convert
a pidl to a path, we should just use the pidl.