winemp3.acm: mpg123_getformat must be called on MPG123_NEW_FORMAT errors for libmpg123 < 1.8.0.
The documentation for libmpg123 clearly state that mpg123_getformat MUST
be called if an error MPG123_NEW_FORMAT is returned, for libmpg123 < 1.8.0.
On versions above that, the call is optional. The patch make sure to call
that function all the time (instead of just when trace is activated) so
that mp3 works in Morrowind and others.