wordpad: Get preview page character boundaries with number of pages.
Rather than making seperate calls to EM_FORMATRANGE to get the number of
pages and the character boundaries, it would be easiest to get both
together.  The time to process EM_FORMATRANGE seems like it would take a
lot longer than the time needed to cache the character page boundaries.