wordpad: Added zoom cursor for print preview.
In print preview you can zoom in by left clicking on one of the pages.  The
zoom cursor is displayed whenever the mouse is over the page to indicate to
the user that they can zoom in.  When fully zoomed in it will zoom out when
this is done again.  When clicking on the second page, it will zoom into
that page.