user32: Do not ask DrawFrameControl() to modify the passed in rectangle.
This patch should fix the regression(s) caused by my recent button painting
patches. The problem was that DrawFrameControl() was modifying the rectangle,
and one of my patches stopped calling DrawFrameControl() in case of ODA_FOCUS
action, so drawing an entire button and only focus started to differ, which
was hidden by the custom clipping (that was removed by my last patch).