wordpad: Fix behaviour of find/replace dialog.
Pressing replace should only replace a selected match, and afterwards it
should find the next match. Searches should always start from the start of
the selection, even if the selection is moved in between searches. Searches
wrap at the end of text, and should keep track of whether it has wrapped
from previous searches to determine if it reached the end position.