wordpad: Remove sized print preview page buffer.
This extra buffer only saves the result of the StretchBlt operatation,
which isn't worth doing since a BitBlt will still be needed from the buffer
to the screen. This causes noticable delay on resizing since a new sized
bitmap needs to be created, the old bitmap needs to be destroyed, and then
both a StretchBlt and a Bitblt is needed to actually repaint the window.