configure: Load wine lower in memory on FreeBSD to leave more space for mmap() and malloc().
If not provided an explicit address, FreeBSD won't mmap any memory
below the executable. In particularly this means any memory below that
is unavailable to dlopen() and malloc().  This fixes the bug where
large WineLib binaries, in particular winetest, failed to load because
there was not enough space left for them and the native libraries they
linked to.  The drawback is that we can no longer load very large
Windows executables (between about 1.5 and 2 GB).